pyproject.toml
[tool.poetry]
name = "sandbox-py"
version = "0.1.0"
description = ""
authors = ["tMinamiii <otherthan.new@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
sqlalchemy = "^1.3.18"
sqlalchemy-stubs = "^0.3"
python-dotenv = "^0.14.0"
elasticsearch_dsl = "^7.2.1"
elasticsearch = "^7.8.0"
flask = "^1.1.2"
flask-httpauth = "^4.1.0"
pytest = "^5.4.3"
pytest-cov = "^2.10.0"
fastapi = "^0.58.1"
pydantic = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8-bugbear = "^20.1.4"
flake8 = "^3.8.3"
ipython = "^7.16.1"
jupyter = "^1.0.0"
mypy = "^0.782"
neovim = "^0.3.1"
pre-commit = "^2.6.0"
pydocstyle = "^5.0.2"
pynvim = "^0.4.1"
pytest = "^5.4.3"
isort = "^5.0.6"

[tool.black]
line-length = 119
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
